Print routes on SIGUSR1
Remove the handler which reads from stdin to trigger the printing of the
peers and routes as debug info. This caused unwanted side effects. The
current functionality will be moved to an API in the future, but for now
it is a sufficiently minor change which solves a couple of open issues.
Also trap SIGTERM for quitting. Functionally this does not change the
behavior of the application, but it might be useful in the future.

Fixes #16
Fixes #17

Signed-off-by: Lee Smet <lee.smet@hotmail.com>