Reduce tokio feature scope
Instead of using "full", we now only use the features we need. This
removes a few downstream dependencies, shortening build times slightly.

This also means the "parking_lot" feature is no longer enabled. It is
generally accepted that the default std::sync::Mutex which is now used
again by tokio is faster for the general case than the parking_lot mutex
implementation.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>