Name the return from Stack::new
Because the returned `Stack` is actually never used, it was explicitly
not assigned (`let _ = ...`). However since the Stack contains the only
reference to the http API, it needs to be kept in scope to prevent that
from shutting down. Instead, we now assign the return value to an
underscored variable name. Semantically this is different, as it keeps
the return value in scope with the usual variable rules instead of
dropping it outright, but still satisfies clippy's unused variable rule.

Fix #140

Signed-off-by: Lee Smet <lee.smet@hotmail.com>