Feature gate messages
Sets up a feature for "message", which enables message related code. The
current implementation is probably not the cleanest, but it gets the job
done. The default feature is set to include the message feature, so
regular builds will not change. But in case it would be useful, the
message code can be removed statically from the binary by building with
--no-default-features.

Closes #130

Signed-off-by: Lee Smet <lee.smet@hotmail.com>