Remove reqwest tls features
Also enable rustls dangerous_configuration feature. This was previously
transitively enabled by the reqwest features.

As a result of this change the CLI won't be able to connect with a
remote endpoint anymore, although it actually couldn't do this in the
first place. It also is a bad idea to expose the API, so not allowing
users to connect with it through the API is fine.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>