Ignore AddrAlreadyInUse errors when joining LL mc
When joining the link local multicast group used for local peer
discovery, there is apparently a condition where the multicast group is
joined but the code is not aware of this (it tries again on the next
iteration). Future calls for this interface will return an
AddrAlreadyInUse error, which we now consider to indicate the interface
is already part of the group, so it is ignored on future calls.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>