Update source table before sending updates
Because of the concurrency primitive used we previously batched update
effects on the source table and applied them once all updates were
processed. This means updates could already be in the process of being
sent, and in case of low latency links, they might even have been
processed already. This could lead to a routing loop because the source
table might not reject the loop update coming back (if it wasn't updated
properly yet).

Fixes #170

Signed-off-by: Lee Smet <lee.smet@hotmail.com>