Update source table only on seqno increase
The previous condition updated the source table unconditionally if the
new seqno was not less than the existing source key seqno. This is not
correct however, as it also updates the source key if the seqno is
the same, thus allowing the source key metric to increase for the same
seqno. This allows for loops to be created, as an update could increase
the metric, after which a loop update could eronously be accepted.

Fixes #176

Signed-off-by: Lee Smet <lee.smet@hotmail.com>