Handle route request for static and missing routes
Previously route requests for specific subnets (i.e. not a wildcard
request for a full table dump) only looked up the selected routes. This
missed the static routes which are not saved in the regular routing
table. Additionally, according to the babel rfc, if no route is found
for the requested subnet, a retraction must be advertised.

Fixes #171

Signed-off-by: Lee Smet <lee.smet@hotmail.com>