Rework router to not use left-right for everything
Previously all fields which could be mutated were stored under a
left-right concurrency primitive. This is neat as it means read-only
access never has to wait. It also means 2 copies of all data are kept.
Since a lot of these fields are not critical for performance, there is
no reason to pay the additional memory cost. On top of this, using
regular locks simplifies the code structure.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>