Save pubkey and shared secret in routing table
Previously a separate table was used for this, which means we wasted
some memory since the same keys are used as the route entries to save
these. Additionally, the keys were not properly cleaned up. Now, keys
are cleaned up when the last route entry for a subnet is removed, and
keys are effectively deduplicated compared to before.

Closes #183

Signed-off-by: Lee Smet <lee.smet@hotmail.com>