Don't take mutable ref of Peer in new hello
Since bumping the seqno can happen with a regular ref, don't require a
mutable reference. This also allows us to just iterate over the Peers
in the router while holding a read lock, avoiding an allocation of a
Vec, and the atomic operations needed to bump the reference counters of
the peers.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>