Don't add LL peer if we already know the IP
Previously we only prevented a peer from being added if the exact
endpoint was already known. For link local discovery, this meant that
there were 2 connections being created, 1 by each end (since they both
discover each other). Now, check when we add an LL connection first if
the link local ip is already connected to with the same protocol. This
effectively means we won't create the outbound connection if the remote
previously established the inbound connection.

Closes #207

Signed-off-by: Lee Smet <lee.smet@hotmail.com>