Unselect route if required after update
After a retraction update it is possible that no route for the subnet is
feasible. We previously did not actually update the routing table in
this case, leaving the old (retracted) route selected, which could cause
a lot of triggered updates being sent around in a multi node
interconnected system.

Fixes #217

Signed-off-by: Lee Smet <lee.smet@hotmail.com>