Fix unselecting selected route after update
Previously we unconditionally unselected the old selected route if no
route was selected after an update. There is however a possible
condition where this triggered both the selected route, which is the one
being updated, to be unselected unconditionally (if the update is
unfeasible), and then be unselected again because it was lost. This
second deselection must not happen.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>