Only log an error once when a static node is not reachable
When a node cannot be reached (e.g. because we're on an IPv4-only
network), don't fill the system logs with messages.
We can log this once as an error and log the rest only when debug
logging is enabled.
In order for this to work, we reset the connection_attempts variable
when we connect successfully to a peer.