Properly send seqno requests when we lose routes
Previously a seqno request was only sent when the selected route
received an unfeasible update. In accordance with the babel rfc, we now
also send seqno requests if we lose a selected route for a subnet
because of route selection (after an update or because we noticed a peer
disappeared).

Closes #233

Signed-off-by: Lee Smet <lee.smet@hotmail.com>