Use cache when forwarding seqno requests
Previously we just tried to find the same route when forwarding a seqno
request. Now we record the forward info in the cache, while also
improving the search. That is, we try to forward to a peer we haven't
used yet (but we don't do a fanout/broadcast search), and also actually
do our very best to forward the request. If we know that didn't work,
use a different peer.

Fixes #169

Signed-off-by: Lee Smet <lee.smet@hotmail.com>