Reduce the seqno cache backing allocation on sweep
The cache seems to stay small when observed on nodes connected to the
network, but still it's good practice to clean up after ourselves. For
now we aggressively shrink as much as possible since the cache is
generally small. If needed (i.e. it is found the cache adds a bunch of
noticeable overhead), we can choose to make this less aggressive in the
future.

Closes #291

Signed-off-by: Lee Smet <lee.smet@hotmail.com>