Immediately remove expired unselected routes
There is no point in keeping an expired route, which is not selected, as
part of the fallback routes with infinite metric. Either there was no
selected route, so it can be removed, or the selected route got
retracted in which case that already caused a retraction to be sent.

Closes #330

Signed-off-by: Lee Smet <lee.smet@hotmail.com>