Improve source table contention
Rewrite the loops which send updates to peers, to create the update
first, apply it to the source table, then loop over the peers. This is
an improvement on the old behavior where updaing the source table was
part of sending to a peer, which lead to a lot of redundant updates to
the source table, and missing entries in case there were only 1 or 2
peers.

Closes #232
Fixes #322

Signed-off-by: Lee Smet <lee.smet@hotmail.com>