Use ArcSwap for subnet route entries
Instead of holding an Arc over the RouteList, hold an
Arc<ArcSwap<RouteList>>. This acts like an AtomicPtr to the RouteList.
This is mainly usefull for updates to existing subnets, since those
won't need to hold a write lock over the writer for the entire routing
table. In effect, this makes most writes executable in parralel to allow
scaling out the update processor to multiple tasks/threads.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>