Route updates for same subnet to the same worker
Change from using a SPMC setup with rendez-vous channel to a setup where
each worker has a dedicated queue, and route the update to a worker
based on the hash of the update subnet. This avoids multiple workers
doing updates to the same subnet. While mulitple updates to a subnet
could be queued, and an update packet was only processed by an update
worker once, multiple consecutive updates to the same subnet could
collide, leading to the RCU firing multiple times for an update. In this
scenario, workers are effectively doing useless work, since it has to be
redone anyway. In the metrics of the public peers, we could see that
despite having more workers, the node processed less updates initially.
Local testing confirmed that multipe updates did conflict, and this
change removes those issues.

It should be noted that this could cause update workers to be starved of
updates to process. We will consider this not an issue, since the law of
large numbers suggests that for a large volume, the hash should be
roughly equal. Also note that the hash is derived on the subnet, which
is not user definieable, so not using a cryptographic hash function here
is not an issue. In case there are a low number of updates in the queues
we also don't have a problem, as low volumes are not the inherent
problem here.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>