Close #529: Make peer connection flush cancellable
Investigation has shown that connection file descriptors could leak in
situations where both peers disconnected with data left on the socket.
While it's not entirely clear how this happened, it seems that the
select was blocked while flushing. To alleviate this, make the flush
itself a branch on the select, and therefore cancellable.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>