Fix #614: Properly reply to local packets for local subnet
Previously packets for the local subnet would trigger a route lookup,
which leads to a no-route entry being generated. Instead, reply to such
packets with an ICMP host unreachable - address uncreachable. This can
be done unilaterally since IF the address would exist on the host, the
packet would not be pushed to the TUN interface by the kernel and thus,
by extension, if such a packet is handled by mycelium, it means there is
no address or local subnet configuration.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>