Rework key lookup to use finite loops instead of recursion
The recursion relied on the fact that the query state times out
relatively quickly. It seems there is at least 1 race condition where
that assumption is false, leading to potentially a stack overflow, or
blocking the system. It's not entirely clear why this happens since
these methods only read the routing table, though it might be that the
caller holds the write handle for the routing table, which prevents the
query timeout expiry in turn causing infinite recursion. Since the query
state only lasts 5 seconds currently, we loop 50 times at most with a
100ms interval. This is for sure not ideal, but for now it solves the
issue. This further emphasises the need to return to synchronous route
discovery, preferably while implementing IS-IS (#656).

Somehow fixes #677

Signed-off-by: Lee Smet <lee.smet@hotmail.com>