Add flushing for peer connections as separate select branch
This restores the old behavior which was introduced to fix an issue
where a peer task could get stuck in a flush call (which was not a
select branch), thus preventing the task from ever exitting, causing a
buildup of tasks (which could keep open OS resources like file
descriptors, causing an eventual resource exhaustion). To do this,
introduce a Connection{Read,Write}Half, and add a `split` method to the
Connection trait. This mimics the behavior of the old code which wrapped
a connection in a framed and then split that entirely.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>